webui-vue/api/requirements.txt
# Z-Image Trainer WebUI API Dependencies
fastapi>=0.110.0
uvicorn>=0.38.0
pydantic>=2.5.0
pillow>=10.0.0
python-multipart>=0.0.6

# WebSocket支持
websockets>=15.0.0
uvloop>=0.22.0
watchfiles>=1.1.0
httptools>=0.7.0

# 核心工具
python-dotenv>=1.0.0
requests>=2.32.0
PyYAML>=6.0.0

